+++ lib/amakanize/series_name.rb
@@ -23,6 +23,7 @@ module Amakanize
           ::Amakanize::Filters::TrailingSeriesNamePayloadDeletionFilter,
           ::Amakanize::Filters::TrailingVolumeNumberDeletionFilter,
           ::Amakanize::Filters::TrailingSeriesNamePayloadDeletionFilter,
+          ::Amakanize::Filters::TrailingSymbolDeletionFilter,
         ]
       end
     end